# meme_backend/src/main/resources/openapi/api.yml
openapi: 3.0.1
info:
  title: Meme backend
  version: 0.0.1
...
servers:
  - url: http://localhost:8080/api
    description: Generated server url
paths:
...
  /v1/account/photo/{photoId}:
    put:
      summary: Updates account photo
      security:
        - BearerAuth:
            - USER_EDIT
      tags:
        - "Account"
      operationId: updateAccountProfilePictureSelf
      parameters:
        - name: photoId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 1
            description: Id of profile photo
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/AccountGetFull'
  /v1/account/{id}/photo/{photoId}:
    put:
      summary: Updates account photo
      security:
        - BearerAuth:
            - MODERATOR_EDIT
      tags:
        - "Account"
      operationId: updateAccountProfilePicture
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 1
            description: Id of account
        - name: photoId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/AccountGetFull'
  /v1/account/password:
    put:
      summary: Updates account password
      security:
        - BearerAuth:
            - USER_EDIT
      tags:
        - "Account"
      operationId: updateAccountPasswordSelf
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountPutPasswordUser'
        required: true
      responses:
        '200':
          description: OK
  /v1/account/{id}/password:
    put:
      summary: Updates account password
      security:
        - BearerAuth:
            - ADMIN_EDIT
      tags:
        - "Account"
      operationId: updateAccountPassword
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 1
            description: Id of account
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountPutPasswordAdmin'
        required: true
      responses:
        '200':
          description: OK
  /v1/account/name/{name}:
    put:
      summary: Updates account password
      security:
        - BearerAuth:
            - USER_EDIT
      tags:
        - "Account"
      operationId: updateAccountNameSelf
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
            example: BigBen
            description: New user name
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/AccountGetFull'
  /v1/account/{id}/name/{name}:
    put:
      summary: Updates account password
      security:
        - BearerAuth:
            - MODERATOR_EDIT
      tags:
        - "Account"
      operationId: updateAccountName
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 1
            description: Id of account
        - name: name
          in: path
          required: true
          schema:
            type: string
            example: BigBen
            description: New user name
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/AccountGetFull'
...
  /v1/account/{id}/ban/{ban}:
    put:
      summary: Updates account banned flag
      security:
        - BearerAuth:
            - MODERATOR_EDIT
      tags:
        - "Account"
      operationId: banAccount
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 1
            description: Account id
        - name: ban
          in: path
          required: true
          schema:
            type: boolean
            example: true
            description: Banned value
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/AccountGetFull'
...
  /v1/account/{id}/role/{newRoleId}:
    put:
      summary: Sets new role for given account
      security:
        - BearerAuth:
            - ADMIN_EDIT
      tags:
        - "Account"
      operationId: changeAccountRole
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 1
            description: Account id
        - name: newRoleId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 1
            description: Account role id
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/AccountRoleGet'
...
  /v1/account/refresh:
    post:
      summary: Checks refresh token and returns new access token
      tags:
        - "Account"
      operationId: refreshJWT
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthenticationRefreshRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/AuthenticationResponse'
  /v1/account/all:
    post:
      summary: Returns paginated list of accounts
      tags:
        - "Account"
      operationId: getAllAccounts
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/AccountPageGet'
  /v1/account/register:
    post:
      summary: Creates new account
      tags:
        - "Account"
      operationId: createAccount
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountPost'
        required: true
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/AccountGetFull'
...
  /v1/account:
    get:
      summary: Returns self account data
      tags:
        - "Account"
      operationId: getSelf
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/AccountGetFull'
    delete:
      summary: Deletes account
      security:
        - BearerAuth:
            - USER_DELETE
      tags:
        - "Account"
      operationId: deleteAccountSelf
      responses:
        '200':
          description: OK
...
  /v1/account/{id}:
    get:
      summary: Returns account by given id
      tags:
        - "Account"
      operationId: getAccount
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            description: Account id
            example: 1
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/AccountGetFull'
    delete:
      summary: Deletes account by given id
      security:
        - BearerAuth:
            - ADMIN_DELETE
      tags:
        - "Account"
      operationId: deleteAccount
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            description: Account id
            example: 1
      responses:
        '200':
          description: OK
  /v1/accountpermission/all:
    get:
      summary: Returns accountpermissions
      security:
        - BearerAuth:
            - ADMIN_READ
      tags:
        - "AccountPermission"
      operationId: getPermissions
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountPermissionGet'
  /v1/accountpermission/{id}:
    get:
      summary: Returns accountpermission by given id
      security:
        - BearerAuth:
            - ADMIN_READ
      tags:
        - "AccountPermission"
      operationId: getPermission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            description: Accountpermission id
            example: 1
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/AccountPermissionGet'
  /v1/account/{id}/role:
    get:
      summary: Returns accountrole by given account id
      security:
        - BearerAuth:
            - ADMIN_READ
      tags:
        - "Account"
      operationId: getAccountRole
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            description: Account id
            example: 1
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/AccountRoleGet'
...
components:
  schemas:
    AccountGetFull:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Id of Account
          example: 2
        name:
          type: string
          description: Name of account
          example: BigBen
        email:
          type: string
          format: email
          description: Email of account
          example: BigBen@gmail.com
        creationDate:
          type: string
          format: date
          description: CreationDate of account
          example: 01.01.2000
        profilePhotoId:
          type: integer
          format: int64
          description: Id of profile photo
          example: 1
        role:
          $ref: '#/components/schemas/AccountRoleGet'
        enabled:
          type: boolean
          description: Determines if account is enabled
          example: true
        banned:
          type: boolean
          description: Determines if account is banned
          example: true
...
    AccountPermissionGet:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Id of permission
          example: 1
        permission:
          type: string
          description: Name of permission
          example: USER_READ
...
    AccountPutPasswordUser:
      type: object
      required:
        - newPassword
        - oldPassword
      properties:
        newPassword:
          type: string
          description: New password of user
          example: SomePassword
        oldPassword:
          type: string
          description: Old password of user
          example: SomePassword
    AccountPutPasswordAdmin:
      type: object
      required:
        - newPassword
      properties:
        newPassword:
          type: string
          description: New password of user
          example: SomePassword
...
    AccountRoleGet:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Account role id
          example: 1
        role:
          type: string
          description: Role name
          example: USER
        permissions:
          type: array
          description: Array of permissions
          items:
            $ref: '#/components/schemas/AccountPermissionGet'
...
    DateRange:
      type: object
      nullable: true
      required:
        - min
        - max
      properties:
        min:
          type: string
          format: date
          description: Minimal date to consider
          example: 01.01.2000
        max:
          type: string
          format: date
          description: Maximal date to consider
          example: 01.01.2000
    LongRange:
      nullable: true
      type: object
      required:
        - min
        - max
      properties:
        min:
          type: integer
          format: int64
          description: Minimal value to consider
          example: 1
        max:
          type: integer
          format: int64
          description: Maximal value to consider
          example: 5
...
    AuthenticationRefreshRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
          description: Refresh token
          example: kjdhaslkfjahdslfkjash
    AuthenticationResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: Access token
          example: kjdhaslkfjahdslfkjash
        refreshToken:
          type: string
          description: Refresh token
          example: kjdhaslkfjahdslfkjash
    AccountFilter:
      type: object
      properties:
        creationDate:
          $ref: '#/components/schemas/DateRange'
        comments:
          $ref: '#/components/schemas/LongRange'
        posts:
          $ref: '#/components/schemas/LongRange'
        upVotes:
          $ref: '#/components/schemas/LongRange'
        downVotes:
          $ref: '#/components/schemas/LongRange'
        accountName:
          type: string
          description: Account name to filter by
          example: BigBen
          nullable: true
        enabled:
          type: boolean
          description: Show enabled accounts
          example: true
          nullable: true
        banned:
          type: boolean
          description: Show banned accounts
          example: true
          nullable: true
    AccountPagingAndSortingRequest:
      type: object
      required:
        - page
        - size
      properties:
        page:
          type: integer
          format: int32
          description: Page to display
          example: 1
        size:
          type: integer
          format: int32
          description: Size of page to display
          example: 1
        sortBy:
          type: string
          description: Sort by field
          example: id
        sortDirection:
          type: string
          description: Sort direction
          example: asc
    AccountRequest:
      type: object
      properties:
        pagingAndSorting:
          $ref: '#/components/schemas/AccountPagingAndSortingRequest'
        filters:
          $ref: '#/components/schemas/AccountFilter'
    AccountPageGet:
      type: object
      properties:
        accounts:
          type: array
          description: Array of accounts
          items:
            $ref: '#/components/schemas/AccountGetFull'
        totalNumberOfPages:
          type: integer
          format: int32
          description: Total number of pages with current filter applied
          example: 1
        totalNumberOfElements:
          type: integer
          format: int64
          description: Total number of elements that can be queried
          example: 1
        filters:
          $ref: '#/components/schemas/AccountFilter'
    AccountPost:
      type: object
      required:
        - name
        - email
        - password
        - profilePictureId
      properties:
        name:
          type: string
          description: Account name
          example: BigBen
        email:
          type: string
          format: email
          description: Account email
          example: BigBen@gmail.com
        password:
          type: string
          format: password
          description: Account password
          example: fkladjdh
        profilePictureId:
          type: integer
          format: int64
          description: Profile picture id
          example: 1
...